Separate command line usage code from crc caluclation.
